# packages/contracts/openapi/dabboba.openapi.yaml
openapi: 3.1.0
info:
  title: DABBOBA API
  version: 0.1.0
  description: Shared user, commerce, and privileged operations contract. Fixture-only prototype behavior is not represented as production persistence.
servers:
  - url: http://127.0.0.1:8788
    description: Isolated local API
...
paths:
...
  /v1/auth/dev-session:
    post:
      tags: [Auth]
      operationId: createDevelopmentSession
      description: Development-only user session. The API refuses this route outside development and test.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 254 }
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCreated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUserSession
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user actor and active session metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSessionState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshUserSession
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Current session revoked and replaced with a newly issued user session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserSessionIssued" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logoutUser
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Current user session revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/admin/auth/login:
    post:
      tags: [Admin, Auth]
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 12, maxLength: 256 }
      responses:
        "201":
          description: Privileged session created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCreated" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/admin/auth/logout:
    post:
      tags: [Admin, Auth]
      operationId: logoutAdmin
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Session revoked. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/admin/me:
    get:
      tags: [Admin]
      operationId: getAdminActor
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current privileged actor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Actor" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the session is invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: The authenticated actor lacks the required role or ownership.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: The visible resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    RateLimited:
      description: Too many attempts.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...
    UserRole: { type: string, enum: [USER, ADMIN, SUPER_ADMIN] }
    UserStatus: { type: string, enum: [ACTIVE, SUSPENDED, BANNED, DELETED] }
...
    Actor:
      type: object
      additionalProperties: false
      required: [userId, email, nickname, role, status, sessionId, permissions]
      properties:
        userId: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        nickname: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        sessionId: { type: string, format: uuid }
        permissions: { type: array, items: { type: string } }
...
    SessionCreated:
      type: object
      additionalProperties: false
      required: [token, expiresAt, actor]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        actor: { $ref: "#/components/schemas/Actor" }
...
    UserActor:
      type: object
      additionalProperties: false
      required: [userId, email, nickname, role, status]
      properties:
        userId: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        nickname: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
    UserSession:
      type: object
      additionalProperties: false
      required: [id, kind, createdAt, lastSeenAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, const: USER }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    UserSessionState:
      type: object
      additionalProperties: false
      required: [actor, session]
      properties:
        actor: { $ref: "#/components/schemas/UserActor" }
        session: { $ref: "#/components/schemas/UserSession" }
    UserSessionIssued:
      type: object
      additionalProperties: false
      required: [token, expiresAt, rotatedFromSessionId, actor, session]
      properties:
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        rotatedFromSessionId: { type: string, format: uuid }
        actor: { $ref: "#/components/schemas/UserActor" }
        session: { $ref: "#/components/schemas/UserSession" }
...